Múltiples Updates
Se añadieron nuevas funcionalidades, mejoras y manejo de errores que contenía la pagina.

+++ Tenri-Backend/backend/database/migrations/2026_05_06_203554_add_avatar_to_users_table.php
@@ -9,14 +9,18 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('avatar')->nullable(); // 👈 Agregamos la columna
+            if (!Schema::hasColumn('users', 'avatar')) {
+                $table->string('avatar')->nullable();
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('avatar');
+            if (Schema::hasColumn('users', 'avatar')) {
+                $table->dropColumn('avatar');
+            }
         });
     }
 };
